requirements.txt
python-telegram-bot
Flask
gunicorn
psycopg2-binary
SQLAlchemy